ENH: Changed TCoordRep default to double to be consistent with Registration framework.  This will allow an instance that only specified TImageType to be compatable with ImageRegistrationMethod.

+++ Code/BasicFilters/itkBSplineInterpolateImageFunction.h
@@ -67,7 +67,7 @@ namespace itk
  */
 template <
   class TImageType, 
-  class TCoordRep = float,
+  class TCoordRep = double,
   class TCoefficientType = double >
 class ITK_EXPORT BSplineInterpolateImageFunction : 
     public InterpolateImageFunction<TImageType,TCoordRep>
